Issue-9705 - Extract range and gen_series functions from functions-array subcrate' s kernels and udf containers
