+++ datafusion/functions-array/src/lib.rs
@@ -35,6 +35,7 @@ mod except;
 mod extract;
 mod kernels;
 mod position;
+mod range;
 mod remove;
 mod replace;
 mod rewrite;
@@ -65,6 +66,8 @@ pub mod expr_fn {
     pub use super::extract::array_slice;
     pub use super::position::array_position;
     pub use super::position::array_positions;
+    pub use super::range::gen_series;
+    pub use super::range::range;
     pub use super::remove::array_remove;
     pub use super::remove::array_remove_all;
     pub use super::remove::array_remove_n;
@@ -86,8 +89,6 @@ pub mod expr_fn {
     pub use super::udf::array_sort;
     pub use super::udf::cardinality;
     pub use super::udf::flatten;
-    pub use super::udf::gen_series;
-    pub use super::udf::range;
 }
 
 /// Registers all enabled packages with a [`FunctionRegistry`]
@@ -95,8 +96,8 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
     let functions: Vec<Arc<ScalarUDF>> = vec![
         string::array_to_string_udf(),
         string::string_to_array_udf(),
-        udf::range_udf(),
-        udf::gen_series_udf(),
+        range::range_udf(),
+        range::gen_series_udf(),
         udf::array_dims_udf(),
         udf::cardinality_udf(),
         udf::array_ndims_udf(),
